Refactor BatchApplication
BatchApplication was divided into BatchApp and BatchJobLauncher.java
BatchJobLauncher do @CacheEvict.
BatchApplication do scheduling.
